requirements.txt
python-decouple
mariadb
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-dht